// src/types/dashboard/vehicleeSchedule.ts/counter.ts
export type Counter = {
  id: number;
  type: string;
  name: string;
  address: string;
  landMark: string | null;
  locationUrl: string | null;
  phone: string | null;
  mobile: string | null;
  fax: string | null;
  email: string | null;
  primaryContactPersonName: string;
  country: string;
  stationId: number;
  status: boolean;
  bookingAllowStatus: string;
  bookingAllowClass: string;
  zone: string | null;
  isSmsSend: boolean;
  createdAt: string;
  updatedAt: string;
};